+++ npm/.flowconfig
@@ -11,4 +11,4 @@ flow-typed
 experimental.const_params=true
 
 [version]
-0.22.1
+>=0.22.1
